# api/app/Docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestao de Propostas API",
        "description": "API REST para gestao de propostas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        }
    ],
    "paths": {
        "/api/v1/clientes": {
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cria um cliente",
                "operationId": "253886e2e0b3fc3fcbb8dcccab29d1c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Detalha um cliente",
                "operationId": "557b6c7a84cc47a6ee329309061a93a5",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ClienteInput": {
                "required": [
                    "nome",
                    "email",
                    "documento"
                ],
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "documento": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Cliente": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "documento": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "integer"
                    },
                    "error": {
                        "type": "integer"
                    },
                    "messages": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            },
...
        }
    },
...
}